[update] 변수 수정

+++ src/main/java/com/example/dhlee/repository/PostRepository.java
@@ -3,5 +3,5 @@ package com.example.dhlee.repository;
 import com.example.dhlee.domain.dto.PostDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<PostDto, Long> {
+public interface PostRepository extends JpaRepository<PostDto, Integer> {
 }
